% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array_ops.R
\name{array_collapse_left}
\alias{array_collapse_left}
\title{array_collapse_left}
\usage{
array_collapse_left(A, B)
}
\arguments{
\item{A}{A 3-D array with shapes nxmxk.}

\item{B}{A matrix with shapes mx1}
}
\description{
array_collapse_left
}
