% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array_ops.R
\name{array_collapse_right}
\alias{array_collapse_right}
\title{array_collapse_right}
\usage{
array_collapse_right(A, B)
}
\arguments{
\item{A}{A 3-D array with shapes nxmxk.}

\item{B}{A matrix with shapes 1xn}
}
\description{
array_collapse_right
}
