% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array_ops.R
\name{array_mult_left}
\alias{array_mult_left}
\title{array_mult_left}
\usage{
array_mult_left(A, B)
}
\arguments{
\item{A}{A 3-D array with shapes nxmxk.}

\item{B}{A matrix with shapes mxl}
}
\description{
array_mult_left
}
