% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array_ops.R
\name{array_mult_right}
\alias{array_mult_right}
\title{array_mult_right}
\usage{
array_mult_right(A, B)
}
\arguments{
\item{A}{A 3-D array with shapes nxmxk.}

\item{B}{A matrix with shapes lxn}
}
\description{
array_mult_right
}
