% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure_helper.R
\name{block_merge}
\alias{block_merge}
\title{block_merge}
\usage{
block_merge(...)
}
\arguments{
\item{...}{<undefined class object> or list: A sequence of block to be merged.}
}
\value{
The merged block as a <undefined class object>.
}
\description{
An auxiliary function to merge blocks.
}
\examples{
level_1=polynomial_block(order=1,values=matrix(c(rep(1,T),rep(0,T)),2,T,byrow=TRUE))
level_2=polynomial_block(order=2,values=matrix(c(rep(0,T),rep(1,T)),2,T,byrow=TRUE))
season_2=harmonic_block(period=20,values=matrix(c(rep(0,T),rep(1,T)),2,T,byrow=TRUE))

final_block=block_merge(level_1,level_2,season_2)
}
