% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure_helper.R
\name{block_mult}
\alias{block_mult}
\title{block_mult}
\usage{
block_mult(block, k)
}
\arguments{
\item{block}{dlm_block: A block to be multiplied.}
}
\value{
The merged block as a dlm_block.
}
\description{
An auxiliary function to merge blocks.
}
\examples{
level <- polynomial_block(alpha = 1, order = 1)

final_block <- block_merge(level, 5)
}
