% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_dirichlet.R
\name{convert_Dirichlet_Normal}
\alias{convert_Dirichlet_Normal}
\title{convert_Dirichlet_Normal}
\usage{
convert_Dirichlet_Normal(ft, Qt, parms)
}
\arguments{
\item{ft}{vector: A vector representing the means from the normal distribution.}

\item{Qt}{matrix: A matrix representing the covariance matrix of the normal distribution.}

\item{parms}{list: A list of extra known parameters of the distribuition. Not used in this function.}
}
\value{
The parameters of the conjugated distribuition of the linear predictor.
}
\description{
DESCRIPTION
}
