% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_normal.R
\name{convert_dummy_post}
\alias{convert_dummy_post}
\title{convert_dummy_post}
\usage{
convert_dummy_post(conj_prior, parms)
}
\arguments{
\item{conj_prior}{list: A vector containing the parameters of the Normal}

\item{parms}{list: A list of extra known parameters of the distribuition. Not used in this function.}
}
\value{
The parameters of the Normal distribuition of the linear predictor.
}
\description{
Calculate the parameters of the Normal that best approximates the given Normal distribuition.
The approximation is the best in the sense that it minimizes the KL divergence from the Normal to the Normal
This funciton is a dummy used for compatibility between kernels.
}
