% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_normal.R
\name{convert_dummy_prior}
\alias{convert_dummy_prior}
\title{convert_dummy_prior}
\usage{
convert_dummy_prior(ft, Qt, parms)
}
\arguments{
\item{ft}{vector: A vector representing the means from the normal distribution.}

\item{Qt}{matrix: A matrix representing the covariance matrix of the normal distribution.}

\item{parms}{list: A list of extra known parameters of the distribuition. Not used in this function.}
}
\value{
The parameters of the conjugated distribuition (which is Normal) of the linear predictor.
}
\description{
Calculate the parameters of the Normal that best approximates the given Normal distribuition.
The approximation is the best in the sense that it minimizes the KL divergence from the Normal to the Normal.
This funciton is a dummy used for compatibility between kernels.
}
