% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson_kernel.R
\name{convert_Gamma_Normal_LB}
\alias{convert_Gamma_Normal_LB}
\title{convert_Gamma_Normal_LB}
\usage{
convert_Gamma_Normal_LB(ft, Qt, parms)
}
\arguments{
\item{ft}{vector: A vector representing the means from the normal distribution.}

\item{Qt}{matrix: A matrix representing the covariance matrix of the normal distribution.}

\item{parms}{list: A list of extra known parameters of the distribuition. Not used in this function.}
}
\value{
The parameters of the posterior distribution.
}
\description{
Calculate the parameters of the Gamma that approximates the given log-Normal distribuition following the approach proposed in ref: Migon.
}
