% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_laplace.R
\name{convert_Laplace_Normal}
\alias{convert_Laplace_Normal}
\title{convert_Laplace_Normal}
\usage{
convert_Laplace_Normal(ft, Qt, parms)
}
\arguments{
\item{ft}{vector: A vector representing the means from the normal distribution.}

\item{Qt}{matrix: A matrix representing the covariance matrix of the normal distribution.}

\item{parms}{list: A list of extra known parameters of the distribution. Not used in this function.}
}
\value{
The parameters of the conjugated distribution of the linear predictor.
}
\description{
This is a dummy function, since, for an Laplace outcome, the conjugated prior is not used.
}
\seealso{
Other {auxiliary functions for a Laplace outcome}: 
\code{\link{Laplace_pred}()},
\code{\link{update_Laplace}()}
}
\concept{{auxiliary functions for a Laplace outcome}}
\keyword{internal}
