% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_multnom.R
\name{convert_Multinom_Normal}
\alias{convert_Multinom_Normal}
\title{convert_Multinom_Normal}
\usage{
convert_Multinom_Normal(ft, Qt, parms = list())
}
\arguments{
\item{ft}{vector: A vector representing the means from the normal distribution.}

\item{Qt}{matrix: A matrix representing the covariance matrix of the normal distribution.}

\item{parms}{list: A list of extra known parameters of the distribuition. Not used in this kernel.}
}
\value{
The parameters of the conjugated distribuition of the linear predictor.
}
\description{
Calculate the parameters of the Dirichlet that best approximates the given log-Normal distribuition.
The approximation is the best in the sense that it minimizes the KL divergence from the log-Normal to the Dirichlet.
}
