% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multnom_kernel.R
\name{convert_Normal_Dir}
\alias{convert_Normal_Dir}
\title{convert_Normal_Dir}
\usage{
convert_Normal_Dir(conj_prior, parms = list())
}
\arguments{
\item{conj_prior}{list: A vector containing the concentration parameters of the Dirichlet.}

\item{parms}{list: A list of extra known parameters of the distribuition. Not used in this kernel.}
}
\value{
The parameters of the Normal distribuition of the linear predictor.
}
\description{
Calculate the parameters of the log-Normal that best approximates the given Dirichlet distribuition.
The approximation is the best in the sense that it minimizes the KL divergence from the Dirichlet to the log-Normal
}
