% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_gamma.R
\name{convert_Normal_FGamma}
\alias{convert_Normal_FGamma}
\title{convert_Normal_FGamma}
\usage{
convert_Normal_FGamma(conj_prior, parms)
}
\arguments{
\item{conj_prior}{list: A vector containing the parameters of the Inverse-Gamma (alpha,beta).}

\item{parms}{list: A list of extra known parameters of the distribuition. Not used in this function.}
}
\value{
The parameters of the Normal distribuition of the linear predictor.
}
\description{
Calculate the parameters of the log-Normal that best approximates the given Inverse-Gamma distribuition.
The approximation is the best in the sense that it minimizes the KL divergence from the Inverse-Gamma to the log-Normal
}
