% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson_kernel.R
\name{convert_Normal_Gamma}
\alias{convert_Normal_Gamma}
\title{convert_Normal_Gamma}
\usage{
convert_Normal_Gamma(conj_prior, parms)
}
\arguments{
\item{conj_prior}{list: A vector containing the parameters of the Gamma (alpha,beta).}

\item{parms}{list: A list of extra known parameters of the distribuition. Not used in this kernel.}
}
\description{
Calculate the parameters of the log-Normal that best approximates the given Gamma distribuition.
The approximation is the best in the sense that it minimizes the KL divergence from the Gamma to the log-Normal
}
