% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma_kernel.R
\name{convert_Normal_IGamma_LB}
\alias{convert_Normal_IGamma_LB}
\title{convert_Normal_IGamma_LB}
\usage{
convert_Normal_IGamma_LB(conj_prior, parms)
}
\arguments{
\item{conj_prior}{list: A vector containing the parameters of the Inverse-Gamma (alpha,beta).}

\item{parms}{list: A list of extra known parameters of the distribuition. Not used in this function.}
}
\description{
Calculate the parameters of the log-Normal that best approximates the given Inverse-Gamma distribuition following the approach proposed in ref: Migon.
}
