% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_gamma_kernel.R
\name{convert_Normal_NG}
\alias{convert_Normal_NG}
\title{convert_Normal_NG}
\usage{
convert_Normal_NG(conj_distr, parms = list())
}
\arguments{
\item{conj_distr}{list: A list containg the parameters of the Normal-Gamma distribuition (mu0,c0,alpha,beta).}

\item{parms}{list: A list of extra known parameters of the distribuition. Not used in this kernel.}
}
\description{
Calculate the parameters of the Normal that best approximates the given Normal-Gamma distribuition.
The approximation is the best in the sense that it minimizes the KL divergence from the Normal-Gamma to the Normal.
In this approach, we suppose that the first entry of the multivariate normal represents the mean of the observed data and the second represent the log variance.
}
