% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_dirichlet.R
\name{Dirichlet}
\alias{Dirichlet}
\title{Dirichlet}
\usage{
Dirichlet(alpha, outcome, offset = outcome^0)
}
\arguments{
\item{alpha}{Vector: A vector of names fir the linear preditor associated with the concentration parameter of the gamma distribuition. The parameter is treated as unknowed and equal to the exponential of the associated linear preditor.}

\item{outcome}{vector: Values of the observed data.}

\item{offset}{vector: The offset at each observation. Must have the same shape as outcome.}
}
\value{
A object of the class dlm_distr
}
\description{
Creates an outcome with Dirichlet distribuition with the chosen parameters.
}
\examples{

T <- 200
a <- 10
b <- 5
data <- rdirichlet(T, c(a, b))

alpha_block <- polynomial_block(alpha = 1) * 2

outcome <- Dirichlet(alpha = c("alpha_1", "alpha_2"), outcome = data)
fitted_data <- fit_model(alpha_block, outcomes = outcome)
summary(fitted_data)
plot(fitted_data)

}
