% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{dlm_sampling}
\alias{dlm_sampling}
\title{dlm_sampling}
\usage{
dlm_sampling(model, sample_size)
}
\arguments{
\item{model}{fitted_dlm: A fitted model from which to sample.}

\item{sample_size}{integer: The number of samples to draw.}
}
\value{
A list containing the following values:
\itemize{
   \item mt Array: An array containing the samples of latent states. Dimensions are n x T x sample_size, where n is the number of latent variable in the model and T is the number of observed values.
   \item ft Array: An array containing the samples of linear predictors. Dimensions are m x T x sample_size, where m is the number of linear predictors in the model and T is the number of observed values.
   \item param Array: An array containing the samples of the parameters of the observational model. Dimensions are k x T x sample_size, where k is the number of parameters in the observational model and T is the number of observed values.
}
}
\description{
dlm_sampling
}
\examples{

T <- 200
mu <- rnorm(T, 0, 0.1)
data <- rnorm(T, cumsum(mu))

level <- polynomial_block(
  mu = 1,
  D = 1 / 0.95
)
variance <- polynomial_block(
  sigma2 = 1
)
# Unknown variance
outcome <- Normal(mu = "mu", sigma2 = "sigma2", outcome = data)

fitted_data <- fit_model(level, variance, outcomes = outcome)
summary(fitted_data)

show_fit(fitted_data, smooth = TRUE)$plot

sample <- dlm_sampling(fitted_data, 2000)
}
