% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{effects.fitted_dlm}
\alias{effects.fitted_dlm}
\title{effects.fitted_dlm}
\usage{
\method{effects}{fitted_dlm}(fitted_dlm, t = fitted_dlm$t, smooth = fitted_dlm$smooth)
}
\description{
effects method for class fitted_dlm
}
