% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{eval_past}
\alias{eval_past}
\title{eval_past}
\usage{
eval_past(model, smooth = FALSE, t_offset = 0, IC_prob = 0.95)
}
\arguments{
\item{model}{<undefined class> or list: The fitted model to be use for evaluation.}

\item{smooth}{bool: The flag indicating if smoothed values should be used. If TRUE, t_offset will not be used.}

\item{t_offset}{positive integer: The relative offset for forecast. Values for time t will be calculated based on the filtered values of time t-t_offset. Will be ignored if smooth is TRUE.}

\item{IC_prob}{Numeric: The credibility level for the I.C. intervals.}
}
\value{
A list containg:
\itemize{
   \item pred Matrix: A matrix with the predictive mean at each time. Dimensions are k x t, where k is the number of outcomes.
   \item var.pred Array: A 3D-array with the predictive covariance matrix at each time. Dimensions are k x k x t, where k is the number of outcomes.
   \item icl.pred Matrix: A matrix with the lower bound of the I.C. based on the credibility given in the arguments. Dimensions are k x t, where k is the number of outcomes.
   \item icu.pred Matrix: A matrix with the upper bound of the I.C. based on the credibility given in the arguments. Dimensions are k x t, where k is the number of outcomes.
}
}
\description{
Evaluates the predictive values for the observed values used to fit the model. Predictions can be made with smoothed values or with filtered values with a time offset.
}
