% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma_kernel.R
\name{Fgamma_pred}
\alias{Fgamma_pred}
\title{Fgamma_pred}
\usage{
Fgamma_pred(conj_param, outcome = NULL, parms = list(), pred_cred = 0.95)
}
\arguments{
\item{conj_param}{List or data.frame: The parameters of the distribuition of the linear predictor.}

\item{outcome}{Vector or matrix (optional): The observed values at the current time. Not used in this function.}

\item{parms}{List: A list of extra parameters for the model. For this function, it must contain the shape parameter phi of the observational model.}

\item{pred_cred}{Numeric: the desired credibility for the credibility interval.}
}
\value{
A list containing the following values:
\itemize{
   \item pred vector/matrix: the mean of the predictive distribuition of a next observation. Same type and shape as the parameter in model.
   \item var.pred vector/matrix: the variance of the predictive distribuition of a next observation. Same type and shape as the parameter in model.
   \item icl.pred vector/matrix: the percentile of 100*((1-pred_cred)/2)% of the predictive distribuition of a next observation. Same type and shape as the parameter in model.
   \item icu.pred vector/matrix: the percentile of 100*(1-(1-pred_cred)/2)% of the predictive distribuition of a next observation. Same type and shape as the parameter in model.
   \item log.like vector: the The log likelihood for the outcome given the conjugated parameters.
}
}
\description{
Calculate the values for the predictive distribuition given the values of the parameter of the distribuition of the linear predictor.
The data is assumed to have Gamma distribuition with unknown shape phi and unknown mean having log-Normal distribuition.
In this scenario, the marginal distribuition of the data is obtained via Monte Carlo.
}
\examples{

params <- data.frame(
  "f1" = c(1:3),
  "f2" = c(3:1),
  "Q11" = rep(1, 3),
  "Q12" = rep(0, 3),
  "Q21" = rep(0, 3),
  "Q22" = rep(1, 3),
)

Fgamma_pred(params)
}
