% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{fit_model}
\alias{fit_model}
\title{fit_model}
\usage{
fit_model(..., data_out, kernel = "Poisson", offset = data_out^0)
}
\arguments{
\item{...}{<undefined class> or list: The structural block of the model.}

\item{data_out}{Matrix: A matrix containing the observed data. Dimensions should be T x k, where T is the time series length and k is the number of outcomes.}

\item{kernel}{String or <undefined class>: The list of functions (or it's name) used to fit the data. Should be choosed based on the distribution of the outcomes.}

\item{offset}{Matrix: A matrix containing the offset value for the data. Dimesions should be the same as y.}
}
\value{
The fitted model (an object of the <undefined class> or list). Contains the values returned in the fit function for the choosed kernel.
}
\description{
Fit a model given it's structure and the observed data. This function can be used for any kernel.
}
\examples{
library(GDLM)

Normal case
T=200
mu=rnorm(T,0,0.1)
y=rnorm(T,cumsum(mu))

level=polynomial_block(order=1,
                       values=matrix(c(rep(1,T),rep(0,T)),2,T,byrow=TRUE),
                       D=1/0.98)
variance=polynomial_block(order=1,
                          values=matrix(c(rep(0,T),rep(1,T)),2,T,byrow=TRUE),
                          D=1/1)

fitted_data=fit_model(level,variance,data_out=matrix(c(y,rep(0,T)),T,2),kernel='Normal')
show_fit(fitted_data,smooth = TRUE)$plot

plot(fitted_data$mt[1,])
lines(fitted_data$mts[1,])

# Poisson case
w=(200/40)*2*pi
y=rpois(T,20*(sin(w*1:T/T)+2))

level=polynomial_block(order=1,values=matrix(1,1,T),D=1/0.95)
season=harmonic_block(period=40,values=matrix(1,1,T),D=1/0.98)

fitted_data=fit_model(level,season,data_out=y,kernel='Poisson')
show_fit(fitted_data,smooth = TRUE)$plot

# Multinomial case
w=(200/40)*2*pi
y1= matrix(rpois(T,20*(sin(w*1:T/T)+2)),T,1)
y2= matrix(rpois(T,1:200/200+1),T,1)
y3= matrix(rpois(T,6),T,1)
y=cbind(y1,y2,y3)

level_1=polynomial_block(order=1,values=matrix(c(rep(1,T),rep(0,T)),2,T,byrow=TRUE))
level_2=polynomial_block(order=2,values=matrix(c(rep(0,T),rep(1,T)),2,T,byrow=TRUE))
season_2=harmonic_block(period=20,values=matrix(c(rep(0,T),rep(1,T)),2,T,byrow=TRUE))


fitted_data=fit_model(level_1,level_2,season_2,data_out=y,kernel='Multinomial')
show_fit(fitted_data,smooth = TRUE)$plot
}
