% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{forecast}
\alias{forecast}
\title{forecast}
\usage{
forecast(
  model,
  t = 1,
  outcome = NULL,
  offset = NULL,
  FF = NULL,
  D = NULL,
  W = NULL,
  plot = TRUE,
  pred_cred = 0.95,
  labels = NULL
)
}
\arguments{
\item{model}{<undefined class> or list: The fitted model to be use for predictions.}

\item{t}{Numeric: Time window for prediction.}

\item{offset}{Matrix or scalar: offset for predictions. Should have dimensions k x t, where k is the number of outcomes of the model. If offset is not specified, the last value observed by the model will be used.}

\item{FF}{Array: A 3D-array containing the regression matrix for each time. It's dimension should be n x k x t, where n is the number of latent variables, k is the number of outcomes in the model. If not specified, the last value given to the model will be used.}

\item{D}{Array: A 3D-array containing the discount factor matrix for each time. It's dimension should be n x n x t, where n is the number of latent variables and T is the time series length. If not specified, the last value given to the model will be used in the first step, and 1 will be use thereafter.}

\item{W}{Array: A 3D-array containing the covariance matrix of the noise for each time. It's dimension should be n x n x t, where n is the number of latent variables and T is the time series length. If not specified, 0 will be used.}

\item{plot}{Bool: A flag indicating if a plot should be produced.}

\item{pred_cred}{Numeric: The credibility level for the I.C. intervals.}

\item{labels}{Vector: A string vector with the names for the series of prediction. If none is given, will use generic names.}
}
\value{
A list containing:
\itemize{
   \item pred Matrix: A matrix with the predictive mean at each time. Dimensions are k x t, where k is the number of outcomes.
   \item var.pred Array: A 3D-array with the predictive covariance matrix at each time. Dimensions are k x k x t, where k is the number of outcomes.
   \item icl.pred Matrix: A matrix with the lower bound of the I.C. based on the credibility given in the arguments. Dimensions are k x t, where k is the number of outcomes.
   \item icu.pred Matrix: A matrix with the upper bound of the I.C. based on the credibility given in the arguments. Dimensions are k x t, where k is the number of outcomes.
   \item at Matrix: A matrix with the values for the linear predictor at each time. Dimensions are k x t, where k is the number of outcomes.
   \item Rt Array: A 3D-array with the covariance of the linear predictor matrix at each time. Dimensions are k x k x t, where k is the number of outcomes.
   \item plot (if so choosed): A plotly object.
}
}
\description{
Makes predictions for t times ahead using the fitted model.
}
\examples{
T <- 200
w <- ((T + 20) / 40) * 2 * pi
y1 <- matrix(rpois((T + 20), 20 * (sin(w * 1:(T + 20) / (T + 20)) + 2)), (T + 20), 1)
y2 <- matrix(rpois((T + 20), 1:(T + 20) / (T + 20) + 1), (T + 20), 1)
y3 <- matrix(rpois((T + 20), 6), (T + 20), 1)
outcome <- cbind(y1, y2, y3)
y_pred <- outcome[T:(T + 20), ]

outcome <- outcome[1:T, ]

level_1 <- polynomial_block(order = 1, values = c(1, 0),by_time = FALSE)
level_2 <- polynomial_block(order = 2, values = c(0, 1),by_time = FALSE)
season_2 <- harmonic_block(period = 20, values = c(0, 1),by_time = FALSE)


fitted_data <- fit_model(level_1, level_2, season_2, outcome = outcome, family = "Multinomial")
show_fit(fitted_data, smooth = TRUE)$plot

forecast(fitted_data, 20, outcome = y_pred)$plot
}
