% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma_kernel.R
\name{gamma_log_like}
\alias{gamma_log_like}
\title{gamma_log_like}
\usage{
gamma_log_like(conj_param, outcome, parms = list())
}
\arguments{
\item{conj_param}{List or data.frame: The parameters of the conjugated distribuition (Gamma) of the linear predictor.}

\item{outcome}{Vector or matrix: The observed values at the current time.}

\item{parms}{List: A list of extra parameters for the model. For this function, it must contain the shape parameter phi of the observational model.}
}
\value{
The predictive log-likelyhood of the observed data.
}
\description{
Calculate the predictive log-likelyhood of the data given the values of the parameter of the conjugated distribuition of the linear predictor.
The data is assumed to have Gamma distribuition with known shape parameter phi and it's mean having distribuition Inverse-Gamma with shape parameter a e rate parameter b.
In this scenario, the marginal distribuition of the data is Beta prime with parameters phi, alpha, beta / phi.
}
\examples{

conj_param <- data.frame(
  "alpha" = c(1:3),
  "beta" = c(3:1)
)

gamma_log_like(conj_param, rgamma(3, c(1:3), c(3:1)), parms = list("phi" = 1))
}
