% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{generic_smoother}
\alias{generic_smoother}
\title{generic_smoother}
\usage{
generic_smoother(mt, Ct, at, Rt, G)
}
\arguments{
\item{mt}{Matrix: A matrix containing the filtered mean of the latent variables at each time. Each row should represent one variable.}

\item{Ct}{Array: A 3D-array representing the filtered covariance matrix of the latent variables at each time. The third dimension should represent the time index.}

\item{at}{Matrix: A matrix containing the one-step-ahead mean of the latent variables at each time based upon the filtered mean. Each row should represent one variable.}

\item{Rt}{Array: A 3D-array representing the one-step-ahead covariance matrix of the latent variables at each time based upon the filtered covariance matrix. The third dimension should represent the time index.}

\item{G}{Array: A 3D-array representing the transition matrix of the model at each time.}
}
\value{
List: The smoothed mean (mts) and covariance (Cts) of the latent variables at each time. Their dimension follows, respectivelly, the dimensions of mt and Ct.
}
\description{
Generic smoother for all models.
}
\examples{
T <- 20

mt <- matrix(c(cumsum(rnorm(T) + 1), rep(1, T)), 2, T, byrow = TRUE)
Ct <- array(diag(c(1, 1)), c(2, 2, T))
G <- array(c(1, 0, 1, 1), c(2, 2,T))
at <- G[,,1] \%*\% mt
Rt <- array(G[,,1]  \%*\% t(G[,,1] ) + diag(c(0.1, 0.1)), c(2, 2, T))

smoothed_values <- generic_smoother(mt, Ct, at, Rt, G)
}
