% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure_helper.R
\name{harmonic_block}
\alias{harmonic_block}
\title{harmonic_block}
\usage{
harmonic_block(
  period,
  values = 1,
  name = "Var_Sazo",
  D = 1,
  m0 = 0,
  C0 = 1,
  W = 0
)
}
\arguments{
\item{W}{}
}
\description{
harmonic_block
}
