% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_laplace.R
\name{Laplace}
\alias{Laplace}
\title{Laplace outcome for kDGLM models}
\usage{
Laplace(mu, outcome, offset = outcome^0)
}
\arguments{
\item{outcome}{Vector: Values of the observed data.}

\item{offset}{Vector: The offset at each observation. Must have the same shape as outcome.}

\item{alpha}{Vector: A vector of names for the linear predictor associated with the mean parameter of the Laplace distribution. The parameter is treated as unknown and equal to the associated linear predictor.}
}
\value{
A object of the class dlm_distr
}
\description{
Creates an outcome with Laplace distribution with the chosen parameters.
}
\details{
For evaluating the posterior parameters, we use a modified version of the method proposed in \insertCite{ArtigokParametrico;textual}{kDGLM}.

For computational efficiency, we also use a Laplace approximations to obtain the first and second moments of the posterior \insertCite{@see @TierneyKadane1 and @TierneyKadane2 }{kDGLM}.

For the details about the implementation see  \insertCite{ArtigoPacote;textual}{kDGLM}.

For the detail about the modification of the method proposed in \insertCite{ArtigokParametrico;textual}{kDGLM}, see \insertCite{ArtigoAltMethod;textual}{kDGLM}.
}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{fit_model}}

Other {auxiliary functions for a creating outcomes}: 
\code{\link{Dirichlet}()},
\code{\link{Gamma}()},
\code{\link{Multinom}()},
\code{\link{Normal}()},
\code{\link{Poisson}()}
}
\concept{{auxiliary functions for a creating outcomes}}
