% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_normal.R
\name{Multi_Normal}
\alias{Multi_Normal}
\title{Normal}
\usage{
Multi_Normal(mu, Sigma, outcome, offset = outcome^0, alt_method = FALSE)
}
\arguments{
\item{mu}{character: Name of the linear preditor associated with the mean parameter of the Normal distribuition. The parameter is treated as unknowed and equal to the associated linear preditor.}

\item{outcome}{vector: Values of the observed data.}

\item{offset}{vector: The offset at each observation. Must have the same shape as outcome.}

\item{tau}{character: Name of the linear preditor associated with the precision parameter of the Normal distribuition. The parameter is treated as unknowed and equal to the exponential of the associated linear preditor. It cannot be specified with sigma or sigma2}

\item{sigma}{character: Name of the linear preditor associated with the scale parameter of the Normal distribuition. The parameter is treated as unknowed and equal to the exponential of the associated linear preditor. It cannot be specified with tau or sigma2.}

\item{sigma2}{character or numeric: Name of the linear preditor associated with the variance parameter of the Normal distribuition. If numeric, this parameter is treated as knowed and equal to the value passed. If a character, the parameter is treated as unknowed and equal to the exponential of the associated linear preditor. It cannot be specified with sigma or tau.}
}
\value{
A object of the class dlm_distr
}
\description{
Normal
}
