% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure_helper.R
\name{multiple_block}
\alias{multiple_block}
\title{multiple_block}
\usage{
multiple_block(ref_block, k, values = ref_block$FF[1, 1, ])
}
\arguments{
\item{ref_block}{<undefined class object> or list: The block to be copied.}

\item{k}{positive integer: The number of outcomes.}

\item{values}{vector: The values of the regressor at each time. If not provided, the first row of the FF array from the ref_block will be used.}
}
\value{
The merged block as a <undefined class object>.
}
\description{
An auxiliary function to repeat the same block for multiple outcomes. Useful if a regressor will be used in all outcomes, but with different effects in each one.
}
\examples{
level_i <- polynomial_block(order = 1, values = 0)
level <- multiple_block(level_i)
}
