% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{multnom_filter}
\alias{multnom_filter}
\title{multnom_filter}
\usage{
multnom_filter(y, m0, C0, FF, G, D, W, offset = c(1, 1, 1), parms = list())
}
\arguments{
\item{y}{Vector: The observed values at time t. Must have length greater than 1.}

\item{m0}{Vector: The prior mean for the latent vector at time t.}

\item{C0}{Matrix: The prior covariance matrix for the latent vector at time t.}

\item{FF}{Matrix: The regression matrix at time t. Should be compatible with the dimensions of m0 and y, i.e., if m0 has dimension n and y has dimension m, FF should be n x m.}

\item{G}{Matrix: The state evolution matrix.}

\item{D}{Matrix: The discount factor matrix at time t.}

\item{W}{Matrix: The noise matrix at time t.}

\item{offset}{Vector: Same dimension as y. A vector contaning the offset at time t.}

\item{parms}{list: a list contaning extra arguments. In this model, extra parameters are not used.}
}
\value{
A list containing:
\itemize{
 \item at: One-step-ahead mean for the latent vectors.
 \item Rt: One-step-ahead covariance matrix for the latent vectors.
 \item ft: One-step-ahead linear predictor.
 \item Qt: One-step-ahead covariance matrix for the linear predictior.
 \item tau: BLANK (see Ref. Raíra).
 \item tau_star: BLANK (see Ref. Raíra).
 \item alpha: BLANK (see Ref. Raíra).
 \item alpha_Star: BLANK (see Ref. Raíra).
 \item mt: The filtered mean of the latent vector at time t.
 \item Ct: The filtered covariance matrix of the latent vector at time t.
 \item y: The observed value at time t.
 \item params: The same as the argument.
 }
}
\description{
Filtering function for the Multinomial model.
}
\examples{
y=c(3,2,5)
m0=log(c(0.5,0.5))
C0=diag(c(1,1))
G=matrix(c(1,0,0,1),2,2)
FF=matrix(c(1,0,0,1),2,2)
D=diag(c(0.1,0.1),2,2)+1
W=diag(c(0,0),2,2)
offset=c(1,1,1)

filtered_data=GDLM::multnom_filter(y,m0,C0,FF,G,D,W,offset)
}
