% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{multnom_fit}
\alias{multnom_fit}
\title{multnom_fit}
\usage{
multnom_fit(y, m0 = 0, C0 = 1, FF, G, D, W, offset, IC_prob = 0.95)
}
\arguments{
\item{y}{Matrix: The observed data. It's dimension shoulb be T x m, where T is the length of the time series and m is the number of outcomes at each time.}

\item{m0}{Vector: The prior mean for the latent vector.}

\item{C0}{Matrix: The prior covariance matrix for the latent vector.}

\item{FF}{Array: A 3D-array containing the regression matrix for each time. It's dimension should be n x m x T, where n is the number of latent variables, m is the number of outcomes in the model and T is the time series length.}

\item{G}{Matrix: The state evolution matrix.}

\item{D}{Array: A 3D-array containing the discount factor matrix for each time. It's dimension should be n x n x T, where n is the number of latent variables and T is the time series length.}

\item{W}{Array: A 3D-array containing the covariance matrix of the noise for each time. It's dimension should be the same as D.}

\item{offset}{Matrix: The offset of the model. It's dimension should be the same as y.}

\item{IC_prob}{Numeric: Deprecated.}
}
\value{
A list containing the following values:
\itemize{
   \item mt Matrix: The filtered mean of the latent variables for each time. Dimensions are n x T.
   \item Ct Array: A 3D-array containing the filtered covariance matrix of the latent variable for each time. Dimensions are n x n x T.
   \item ft Matrix: The one-step-ahead linear predictor for each time. Dimensions are m x T.
   \item Qt Array: A 3D-array containing the one-step-ahead covariance matrix for the linear predictor for each time. Dimensions are m x T.
   \item alpha Matrix: BLANK (see Ref. Raíra)
   \item alpha_star Matrix: BLANK (see Ref. Raíra)
   \item tau Matrix: BLANK (see Ref. Raíra)
   \item tau_star Matrix: BLANK (see Ref. Raíra)
   \item FF Array: The same as the argument (same values).
   \item G Matrix: The same as the argument (same values).
   \item D Array: The same as the argument (same values).
   \item W Array: The same as the argument (same values).
   \item pred Matrix: The one-step-ahead predictions for each time. Dimensions are m x T.
   \item var.pred Matrix: The variance for the one-step-ahead predictions for each time. Dimensions are m x T. Note that, in the multivariate Poisson case, the series are supossed independent, so, in particular, they are uncorrelated.
   \item icl.pred Matrix: The lower credibility interval for the prediction at each time. Dimensions are m x T.
   \item icu.pred Matrix: The upper credibility interval for the prediction at each time. Dimensions are m x T.
   \item mts Matrix: The smoothed mean of the latent variables for each time. Dimensions are n x T.
   \item Cts Array: A 3D-array containing the smoothed covariance matrix of the latent variable for each time. Dimensions are n x n x T.
   \item IC_prob Numeric: Deprecated
   \item offset Vector: The same as the argument (same values).
   \item log_offset Vector: The log offset.
   \item data_out Matrix: The same as the argument y (same values).
}
}
\description{
Fit the multinomial model given the observed value and the model parameters.
}
\examples{
# Not ideal way: should use fit_model function.
T=200
w=(200/40)*2*pi
y1= matrix(rpois(T,20*(sin(w*1:T/T)+2)),T,1)
y2= matrix(rpois(T,1:200/200+1),T,1)
y3= matrix(rpois(T,6),T,1)
y=cbind(y1,y2,y3)
m0=c(0,0,0,0,0)
C0=diag(5)
G1=as.matrix(Matrix::bdiag(1,matrix(c(cos(w),sin(w),-sin(w),cos(w)),2,2)))
G2=matrix(c(1,0,1,1),2,2)
G=as.matrix(Matrix::bdiag(G1,G2))
FF=array(matrix(c(1,1,0,0,0,0,0,0,1,0),5,2),c(5,2,T))
D=array(diag(c(0.1,0,0,0.1,0)),c(5,5,T))+1
W=array(diag(c(0,0,0,0,0)),c(5,5,T))
offset=matrix(1,T,3)

fitted_model=GDLM::multnom_fit(y=y,m0 = m0, C0 = C0, FF=FF,G=G,D=D,W=W, offset=offset, IC_prob=0.95)

plot(y1,col='red',ylim=c(0,max(y)*1.2))
points(y2,col='green')
points(y3,col='blue')
lines(fitted_model$pred[1,],col='red')
lines(fitted_model$pred[2,],col='green')
lines(fitted_model$pred[3,],col='blue')
}
