% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multnom_kernel.R
\name{multnom_log_like}
\alias{multnom_log_like}
\title{multnom_log_like}
\usage{
multnom_log_like(conj_param, outcome, parms = list())
}
\arguments{
\item{conj_param}{List or data.frame: The parameters of the conjugated distribuition (Dirichlet) of the linear predictor.}

\item{outcome}{Vector or matrix: The observed values at the current time.}

\item{parms}{List (optional): A list of extra parameters for the model. Not used in this function.}
}
\value{
The predictive log-likelyhood of the observed data.
}
\description{
Calculate the predictive log-likelyhood of the data given the values of the parameter of the conjugated distribuition of the linear predictor.
The data is assumed to have Multinomial distribuition with known number of trial N and the probability vector having distribuition Dirichlet with parameters alpha_i.
In this scenario, the marginal distribuition of the data is Dirichlet-Multinomial with parameters N and alpha_i.
}
\examples{

conj_param <- matrix(c(1:15), 5, 3, byrow = TRUE)

multnom_log_like(conj_param, matrix(5, 5, 3))
}
