% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{multnom_pred}
\alias{multnom_pred}
\title{multnom_pred}
\usage{
multnom_pred(model, IC_prob = 0.95)
}
\arguments{
\item{model}{List: contaning the parameters of the conjugated distribuition for the linear predictor (may be based on the prior, filtered or smoothed distribuition).}

\item{IC_prob}{Numeric: the desired credibility for the credibility interval}
}
\value{
A list containing the following values:
\itemize{
   \item pred vector: the mean of the predictive distribuition of a next observation. Same type and shape as the parameter in model.
   \item var.pred vector: the variance of the predictive distribuition of a next observation. Same type and shape as the parameter in model.
   \item icl.pred vector: the percentile of 100*((1-IC_prob)/2)% of the predictive distribuition of a next observation. Same type and shape as the parameter in model.
   \item icu.pred vector: the percentile of 100*(1-(1-IC_prob)/2)% of the predictive distribuition of a next observation. Same type and shape as the parameter in model.
}
}
\description{
This function is a temporary placeholder.
Calculate the values for the predictive distribuition given the values of the parameter of the conjugated distribuition of the linear predictor.
PLACE PREDICTIVE DISTRIBUITION.
}
\examples{
# MUST BE MADE.
# A fitted model shoulb be used as argument, but you can also pass only the parameter themselves.

model=list(
'ft'=log(c(1:3)),
'Qt'=diag(3),
'y'=c(3,2,5)
)

multnom_pred(model)
}
