% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_kernel.R
\name{normal_filter}
\alias{normal_filter}
\title{normal_filter}
\usage{
normal_filter(outcome, m0, C0, FF, G, D, W, offset = 1, parms = list())
}
\arguments{
\item{outcome}{Vector: The observed values at time t. Must have length greater than 1.}

\item{m0}{Vector: The prior mean for the latent vector at time t.}

\item{C0}{Matrix: The prior covariance matrix for the latent vector at time t.}

\item{FF}{Matrix: The regression matrix at time t. Should be compatible with the dimensions of m0 and outcome, i.e., if m0 has dimension n and outcome has dimension m, FF should be n x m.}

\item{G}{Matrix: The state evolution matrix.}

\item{D}{Matrix: The discount factor matrix at time t.}

\item{W}{Matrix: The noise matrix at time t.}

\item{offset}{Vector: Same dimension as outcome. A vector contaning the offset at time t.}

\item{parms}{list: a list contaning extra arguments. In this model, sigma2 (covariance matrix for the noise associated withy) should be passed.}
}
\value{
A list containing:
\itemize{
 \item at: One-step-ahead mean for the latent vectors.
 \item Rt: One-step-ahead covariance matrix for the latent vectors.
 \item ft: One-step-ahead linear predictor.
 \item Qt: One-step-ahead covariance matrix for the linear predictior.
 \item mt: The filtered mean of the latent vector at time t.
 \item Ct: The filtered covariance matrix of the latent vector at time t.
 \item outcome: The observed value at time t.
 \item params: The same as the argument.
 }
}
\description{
Filtering function for the Normal model with known variance.
}
\examples{
outcome <- c(rnorm(1, 0, 1), 0)
m0 <- c(2, -log(1.1))
C0 <- diag(2)
G <- matrix(c(1, 0, 0, 1), 2, 2)
FF <- matrix(c(1, 0, 0, 1), 2, 2)
D <- diag(c(0.1, 0.1), 2, 2) + 1
W <- diag(c(0, 0), 2, 2)
offset <- c(0)
filtered_data <- GDLM::normal_filter(outcome, m0, C0, FF, G, D, W, offset, parms = list("sigma2" = 1))
}
