% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{normal_fit}
\alias{normal_fit}
\title{Title}
\usage{
normal_fit(
  y,
  m0 = 0,
  C0 = 1,
  FF,
  G,
  D,
  W,
  offset,
  IC_prob = 0.95,
  parms = list()
)
}
\arguments{
\item{parms}{}
}
\description{
Title
}
