% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_gamma_kernel.R
\name{normal_gamma_log_like}
\alias{normal_gamma_log_like}
\title{normal_gamma_log_like}
\usage{
normal_gamma_log_like(conj_param, outcome, parms = list())
}
\arguments{
\item{conj_param}{List or data.frame: The parameters of the conjugated distribuition (Normal-Gamma) of the linear predictor.}

\item{outcome}{Vector or matrix: The observed values at the current time.}

\item{parms}{List (optional): A list of extra parameters for the model. Not used in this function.}
}
\value{
The predictive log-likelyhood of the observed data.
}
\description{
Calculate the predictive log-likelyhood of the data given the values of the parameter of the conjugated distribuition of the linear predictor.
The data is assumed to have Normal distribuition with it's mean and precision having joint distribuition Normal-Gamma with parameters mu0, c0, alpha, beta.
In this scenario, the marginal distribuition of the data is t-student with 2*alpha degrees of freedoom, mu0 as the location parameter and (beta/alpha)*(1+1/c0) as the scale parameter.
}
\examples{

conj_param <- data.frame(
  "mu0" = rep(0, 3),
  "c0" = rep(1, 3),
  "alpha" = rep(3, 3),
  "beta" = rep(3, 3)
)

normal_gamma_log_like(conj_param, rnorm(3))
}
