% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_kernel.R
\name{normal_gamma_pred}
\alias{normal_gamma_pred}
\title{normal_gamma_pred}
\usage{
normal_gamma_pred(conj_param, outcome = NULL, parms = list(), pred_cred = 0.95)
}
\arguments{
\item{conj_param}{List or data.frame: The parameters of the conjugated distribuition (Normal-Gamma) of the linear predictor.}

\item{outcome}{Vector or matrix (optional): The observed values at the current time.}

\item{parms}{List (optional): A list of extra parameters for the model. Not used in this function.}

\item{pred_cred}{Numeric: the desired credibility for the credibility interval.}
}
\value{
A list containing the following values:
\itemize{
   \item pred vector/matrix: the mean of the predictive distribuition of a next observation. Same type and shape as the parameter in model.
   \item var.pred vector/matrix: the variance of the predictive distribuition of a next observation. Same type and shape as the parameter in model.
   \item icl.pred vector/matrix: the percentile of 100*((1-pred_cred)/2)% of the predictive distribuition of a next observation. Same type and shape as the parameter in model.
   \item icu.pred vector/matrix: the percentile of 100*(1-(1-pred_cred)/2)% of the predictive distribuition of a next observation. Same type and shape as the parameter in model.
   \item log.like vector: the The log likelihood for the outcome given the conjugated parameters.
}
}
\description{
Calculate the values for the predictive distribuition given the values of the parameter of the conjugated distribuition of the linear predictor.
The data is assumed to have Normal distribuition with it's mean and precision having joint distribuition Normal-Gamma with parameters mu0, c0, alpha, beta.
In this scenario, the marginal distribuition of the data is t-student with 2*alpha degrees of freedoom, mu0 as the location parameter and (beta/alpha)*(1+1/c0) as the scale parameter.
}
\examples{

conj_param <- data.frame(
  "mu0" = rep(0, 3),
  "c0" = rep(1, 3),
  "alpha" = rep(3, 3),
  "beta" = rep(3, 3)
)

normal_gamma_pred(conj_param)
}
