% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_kernel.R
\name{normal_pred}
\alias{normal_pred}
\title{normal_pred}
\usage{
normal_pred(conj_param, outcome, parms = list(), pred_cred = 0.95)
}
\arguments{
\item{pred_cred}{Numeric: the desired credibility for the credibility interval.}

\item{model}{List: contaning the parameters of the conjugated distribuition for the linear predictor (may be based on the prior, filtered or smoothed distribuition).}
}
\value{
A list containing the following values:
\itemize{
   \item pred vector/matrix: the mean of the predictive distribuition of a next observation. Same type and shape as the parameter in model.
   \item var.pred vector/matrix: the variance of the predictive distribuition of a next observation. Same type and shape as the parameter in model.
   \item icl.pred vector/matrix: the percentile of 100*((1-pred_cred)/2)% of the predictive distribuition of a next observation. Same type and shape as the parameter in model.
   \item icu.pred vector/matrix: the percentile of 100*(1-(1-pred_cred)/2)% of the predictive distribuition of a next observation. Same type and shape as the parameter in model.
}
}
\description{
Calculate the values for the predictive distribuition given the values of the parameter of the conjugated distribuition of the linear predictor.
}
\examples{
# A fitted model shoulb be used as argument, but you can also pass only the parameter themselves.

model <- list(
  "tau0" = 0.5,
  "tau1" = -1 / 2,
  "tau2" = 0,
  "tau3" = -2.0,
  "Qt" = matrix(1, 1, 1),
  "outcome" = 0
)

normal_pred(model)
}
