% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.fitted_dlm}
\alias{plot.fitted_dlm}
\title{plot.fitted_dlm}
\usage{
\method{plot}{fitted_dlm}(model, pred_cred = 0.95, smooth = model$smooth, plotly = TRUE, h = 0)
}
\description{
plot method for class fitted_dlm
}
