% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helper.R
\name{plot_lat_var}
\alias{plot_lat_var}
\title{plot_lat_var}
\usage{
plot_lat_var(
  model,
  var,
  smooth = TRUE,
  cut_off = 10,
  IC_prob = 0.95,
  dinamic = TRUE,
  tranform_y = function(y) {
     y
 }
)
}
\arguments{
\item{tranform_y}{}
}
\description{
plot_lat_var
}
