% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson_kernel.R
\name{Poisson}
\alias{Poisson}
\title{Poisson}
\usage{
Poisson(lambda, outcome, offset = outcome^0)
}
\arguments{
\item{lambda}{character: Name of the linear preditor associated with the rate (mean) parameter of the Poisson distribuition. The parameter is treated as unknowed and equal to the exponential of the associated linear preditor.}

\item{outcome}{vector: Values of the observed data.}

\item{offset}{vector: The offset at each observation. Must have the same shape as outcome.}
}
\value{
A object of the class dlm_distr
}
\description{
Creates an outcome with Poisson distribuition with the chosen parameter.
}
\examples{

# Poisson case
T <- 200
w <- (200 / 40) * 2 * pi
data <- rpois(T, 20 * (sin(w * 1:T / T) + 2))

level <- polynomial_block(rate = 1, D = 1 / 0.95)
season <- harmonic_block(rate = 1, period = 40, D = 1 / 0.98)

outcome <- Poisson(lambda = "rate", outcome = data)

fitted_data <- fit_model(level, season, outcomes = outcome)
summary(fitted_data)

show_fit(fitted_data, smooth = TRUE)$plot
}
