% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{poisson_filter}
\alias{poisson_filter}
\title{poisson_filter}
\usage{
poisson_filter(y, m0, C0, FF, G, D, W, offset)
}
\arguments{
\item{y}{Vector: The observed value at time t.}

\item{m0}{Vector: The prior mean for the latent vector at time t.}

\item{C0}{Matrix: The prior covariance matrix for the latent vector at time t.}

\item{FF}{Matrix: The regression matrix at time t. Should be compatible with the dimensions of m0 and y, i.e., if m0 has dimension n and y has dimension m, FF should be n x m.}

\item{G}{Matrix: The state evolution matrix.}

\item{D}{Matrix: The discount factor matrix at time t.}

\item{W}{Matrix: The noise matrix at time t.}

\item{pop}{Vector: Same dimension as y. A vector contaning the offset at time t.}
}
\value{
A list containing:
- at: One-step-ahead mean for the latent vectors.
- Rt: One-step-ahead covariance matrix for the latent vectors.
- ft: One-step-ahead linear predictor.
- qt: One-step-ahead covariance matrix for the linear predictior.
- a: The \alpha parameter of the compatibilized gamma prior (see Ref. Raíra).
- b: The \beta parameter of the compatibilized gamma prior (see Ref. Raíra).
- a.post: The \alpha parameter of the gamma posteirior (see Ref. Raíra).
- b.post: The \beta parameter of the gamma posteirior (see Ref. Raíra).
- mt: The filtered mean of the latent vector at time t.
- Ct: The filtered covariance matrix of the latent vector at time t.
- y: The observed value at time t.
}
\description{
Filtering function for the Poisson model.
}
