% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson_kernel.R
\name{poisson_lb_filter}
\alias{poisson_lb_filter}
\title{poisson_lb_filter}
\usage{
poisson_lb_filter(outcome, m0, C0, FF, G, D, W, offset = 1, parms = list())
}
\arguments{
\item{outcome}{Vector: The observed value at time t.}

\item{m0}{Vector: The prior mean for the latent vector at time t.}

\item{C0}{Matrix: The prior covariance matrix for the latent vector at time t.}

\item{FF}{Matrix: The regression matrix at time t. Should be compatible with the dimensions of m0 and outcome, i.e., if m0 has dimension n and outcome has dimension m, FF should be n x m.}

\item{G}{Matrix: The state evolution matrix.}

\item{D}{Matrix: The discount factor matrix at time t.}

\item{W}{Matrix: The noise matrix at time t.}

\item{offset}{Vector: Same dimension as outcome. A vector contaning the offset at time t.}

\item{parms}{list: a list contaning extra arguments. In this model, extra parameters are not used.}
}
\value{
A list containing:
\itemize{
 \item at: One-step-ahead mean for the latent vectors.
 \item Rt: One-step-ahead covariance matrix for the latent vectors.
 \item ft: One-step-ahead linear predictor.
 \item qt: One-step-ahead covariance matrix for the linear predictior.
 \item a: The alpha parameter of the compatibilized gamma prior (see Ref. Raíra).
 \item b: The beta parameter of the compatibilized gamma prior (see Ref. Raíra).
 \item a.post: The alpha parameter of the gamma posteirior (see Ref. Raíra).
 \item b.post: The beta parameter of the gamma posteirior (see Ref. Raíra).
 \item mt: The filtered mean of the latent vector at time t.
 \item Ct: The filtered covariance matrix of the latent vector at time t.
 \item outcome: The observed value at time t.
 \item parms: The same as the argument.
 }
}
\description{
Filtering function for the Poisson model using Linear Bayes.
}
\examples{
outcome <- 5
m0 <- c(4, 1)
C0 <- diag(c(1, 1))
G <- matrix(c(1, 0, 1, 1), 2, 2)
FF <- matrix(c(1, 0), 2, 1)
D <- diag(c(0.1, 0.1), 2, 2) + 1
W <- diag(c(0, 0), 2, 2)
offset <- 1

filtered_data <- poisson_filter(outcome, m0, C0, FF, G, D, W, offset)
}
