% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson_kernel.R
\name{poisson_log_like}
\alias{poisson_log_like}
\title{poisson_log_like}
\usage{
poisson_log_like(conj_param, outcome, parms = list())
}
\arguments{
\item{conj_param}{List or data.frame: The parameters of the conjugated distribuition (Gamma) of the linear predictor.}

\item{outcome}{Vector or matrix: The observed values at the current time.}

\item{parms}{List (optional): A list of extra parameters for the model. Not used in this function.}
}
\value{
The predictive log-likelyhood of the observed data.
}
\description{
Calculate the predictive log-likelyhood of the data given the values of the parameter of the conjugated distribuition of the linear predictor.
The data is assumed to have Poisson distribuition with it's mean having distribuition Gamma with shape parameter a e rate parameter b.
In this scenario, the marginal distribuition of the data is Negative Binomial with a as the dispersion parameter and b/(b+1) as the probability.
}
\examples{

conj_param <- data.frame(
  "a" = c(1:3),
  "b" = c(3:1)
)

poisson_log_like(conj_param, rpois(3, 1))
}
