% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure_helper.R
\name{polynomial_block}
\alias{polynomial_block}
\title{polynomial_block}
\usage{
polynomial_block(
  order,
  values = 1,
  name = "Var_Poly",
  D = 1,
  m0 = 0,
  C0 = 1,
  W = 0
)
}
\arguments{
\item{W}{}
}
\description{
polynomial_block
}
