% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{predict}
\alias{predict}
\title{predict}
\usage{
predict(
  model,
  t = 1,
  offset = NULL,
  FF = NULL,
  D = NULL,
  W = NULL,
  plot = TRUE,
  IC_prob = 0.95,
  labels = NULL
)
}
\arguments{
\item{model}{<undefined class> or list: The fitted model to be use for predictions.}

\item{t}{Numeric: Time window for prediction.}

\item{offset}{Matrix: offset for predictions. Should have dimensions k x t, where k is the number of outcomes of the model. If offset is not specified, the last value observed by the model will be used.}

\item{FF}{Array: A 3D-array containing the regression matrix for each time. It's dimension should be n x k x t, where n is the number of latent variables, k is the number of outcomes in the model. If not specified, the last value given to the model will be used.}

\item{D}{Array: A 3D-array containing the discount factor matrix for each time. It's dimension should be n x n x t, where n is the number of latent variables and T is the time series length. If not specified, the last value given to the model will be used in the first step, and 1 will be use thereafter.}

\item{W}{Array: A 3D-array containing the covariance matrix of the noise for each time. It's dimension should be n x n x t, where n is the number of latent variables and T is the time series length. If not specified, 0 will be used.}

\item{plot}{Bool: A flag indicating if a plot should be produced.}

\item{IC_prob}{Numeric: The credibility level for the I.C. intervals.}

\item{labels}{Vector: A string vector with the names for the series of prediction. If none is given, will use generic names.}
}
\value{
A list containing:
\itemize{
   \item pred Matrix: A matrix with the predictive mean at each time. Dimensions are k x t, where k is the number of outcomes.
   \item var.pred Array: A 3D-array with the predictive covariance matrix at each time. Dimensions are k x k x t, where k is the number of outcomes.
   \item icl.pred Matrix: A matrix with the lower bound of the I.C. based on the credibility given in the arguments. Dimensions are k x t, where k is the number of outcomes.
   \item icu.pred Matrix: A matrix with the upper bound of the I.C. based on the credibility given in the arguments. Dimensions are k x t, where k is the number of outcomes.
   \item at Matrix: A matrix with the values for the linear predictor at each time. Dimensions are k x t, where k is the number of outcomes.
   \item Rt Array: A 3D-array with the covariance of the linear predictor matrix at each time. Dimensions are k x k x t, where k is the number of outcomes.
   \item plot (if so choosed): A plotly object.
}
}
\description{
Makes predictions for t times ahead using the fitted model.
}
