% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.fitted_dlm}
\alias{print.fitted_dlm}
\title{print.fitted_dlm}
\usage{
\method{print}{fitted_dlm}(
  fitted_dlm,
  t = fitted_dlm$t,
  smooth = TRUE,
  metric_cutoff = round(fitted_dlm$t/10)
)
}
\description{
print method for class fitted_dlm
}
