% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helper.R
\name{show_fit}
\alias{show_fit}
\title{show_fit}
\usage{
show_fit(
  model,
  pred_cred = 0.95,
  smooth = TRUE,
  dynamic_plot = TRUE,
  t_offset = 0
)
}
\arguments{
\item{model}{fitted_dlm: A fitted GDLM model.}

\item{pred_cred}{Numeric: The credibility value for the credibility interval.}

\item{smooth}{Bool: A flag indicating if the smoothed should be used. If false, the filtered distribuition will be used.}

\item{dynamic_plot}{Bool: A flag indicating if the created plot should be dynamic.}

\item{t_offset}{Integer: A integer with the amount of steps ahead should be used for prediciton. Only used if smooth is false.}
}
\value{
A list containg:
\itemize{
   \item data tibble object: A data frame containing the observations, predictions and credibility intervals at each time.
   \item plot ggplot or plotly object: A plot showing the predictive mean and credibility interval with the observed data.
}
}
\description{
Calculate the preditive mean and some quantiles for the observed data and show a plot.
}
\examples{
T <- 200
w <- (200 / 40) * 2 * pi
data <- rpois(T, 20 * (sin(w * 1:T / T) + 2))

level <- polynomial_block(rate = 1, D = 1 / 0.95)
season <- harmonic_block(rate = 1, period = 40, D = 1 / 0.98)

outcome <- Poisson(lambda = "rate", outcome = data)

fitted_data <- fit_model(level, season, outcomes = outcome)
summary(fitted_data)

show_fit(fitted_data, smooth = TRUE)$plot
}
