% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.fitted_dlm}
\alias{summary.fitted_dlm}
\title{summary.fitted_dlm}
\usage{
\method{summary}{fitted_dlm}(
  fitted_dlm,
  t = fitted_dlm$t,
  smooth = TRUE,
  metric_cutoff = round(fitted_dlm$t/10)
)
}
\description{
summary method for class fitted_dlm
}
