% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_multnom.R
\name{system_multinom}
\alias{system_multinom}
\title{system_multinom}
\usage{
system_multinom(x, parms)
}
\arguments{
\item{x}{vector: current tau values.}

\item{parms}{list: auxiliary values for the system.}
}
\value{
a vector with the values of the system (see Ref. Raíra).
}
\description{
Evaluate the compatibilizing equation for the multinomial model (see Ref. Raíra).
}
