% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multnom_kernel.R
\name{update_Dir}
\alias{update_Dir}
\title{update_Dir}
\usage{
update_Dir(conj_prior, y, parms = list())
}
\arguments{
\item{conj_prior}{list: A vector containing the concentration parameters of the Dirichlet.}

\item{y}{vector: A vector containing the observations.}

\item{parms}{list: A list of extra known parameters of the distribuition. Not used in this kernel.}
}
\description{
Calculate posterior parameter for the Dirichlet, assuming that the observed values came from a Multinomial model from which the number of trials is known and the prior distribuition for the probabilities of each category have joint distribuition Dirichlet.
}
