% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_gamma_kernel.R
\name{update_FGamma}
\alias{update_FGamma}
\title{update_FGamma}
\usage{
update_FGamma(conj_prior, y, parms)
}
\arguments{
\item{conj_prior}{list: A vector containing the parameters of the Inverse-Gamma (alpha,beta).}

\item{y}{vector: A vector containing the observations.}

\item{parms}{list: A list of extra known parameters of the distribuition. For this kernel, parms should containg the shape parameter (phi) for the observational gamma model.}
}
\value{
The parameters of the posterior distribution.
}
\description{
Calculate posterior parameter for the Inverser-Gamma, assuming that the observed values came from a Gamma model from which the shape parameter (phi) is known and the mean (mu) have prior distribuition Inverse-Gamma.
}
