% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson_kernel.R
\name{update_Gamma}
\alias{update_Gamma}
\title{update_Gamma}
\usage{
update_Gamma(conj_prior, y, parms)
}
\arguments{
\item{conj_prior}{list: A vector containing the parameters of the Gamma (alpha,beta).}

\item{y}{vector: A vector containing the observations.}

\item{parms}{list: A list of extra known parameters of the distribuition. Not used in this kernel.}
}
\description{
Calculate posterior parameter for the Gamma, assuming that the observed values came from a Poisson model from which the rate parameter (lambda) have prior distribuition Gamma.
}
