% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_gamma.R
\name{update_Gamma_alt}
\alias{update_Gamma_alt}
\title{update_Gamma_alt}
\usage{
update_Gamma_alt(conj_prior, ft, Qt, y, parms)
}
\arguments{
\item{conj_prior}{list: A vector containing the parameters of the Inverse-Gamma (alpha,beta). Not used in the alternative method.}

\item{ft}{vector: A vector representing the means from the normal distribution.}

\item{Qt}{matrix: A matrix representing the covariance matrix of the normal distribution.}

\item{y}{vector: A vector containing the observations.}

\item{parms}{list: A list of extra known parameters of the distribution. For this kernel, parms should containg the shape parameter (phi) for the observational gamma model.}
}
\value{
The parameters of the posterior distribution.
}
\description{
Calculate posterior parameter for the Inverse-Gamma, assuming that the observed values came from a Gamma model from which the shape parameter (phi) is known and the mean (mu) have prior distribuition Inverse-Gamma.
}
