% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_multnom.R
\name{update_Multinom_alt}
\alias{update_Multinom_alt}
\title{update_Multinom_alt}
\usage{
update_Multinom_alt(conj_prior, ft, Qt, y, parms = list())
}
\arguments{
\item{conj_prior}{list: A vector containing the concentration parameters of the Dirichlet. Not used in the alternative method.}

\item{ft}{vector: A vector representing the means from the normal distribution.}

\item{Qt}{matrix: A matrix representing the covariance matrix of the normal distribution.}

\item{y}{vector: A vector containing the observations.}

\item{parms}{list: A list of extra known parameters of the distribuition. Not used in this kernel.}
}
\value{
The parameters of the posterior distribution.
}
\description{
Calculate posterior parameter for the Dirichlet, assuming that the observed values came from a Multinomial model from which the number of trials is known and the prior distribuition for the probabilities of each category have joint distribuition Dirichlet.
}
