% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_gamma_kernel.R
\name{update_NG}
\alias{update_NG}
\title{update_NG}
\usage{
update_NG(conj_prior, y, parms = list())
}
\arguments{
\item{conj_prior}{list: A vector containing the parameters of the Normal-Gamma (mu0,c0,alpha,beta).}

\item{y}{vector: A vector containing the observations.}

\item{parms}{list: A list of extra known parameters of the distribuition. Not used in this kernel.}
}
\description{
Calculate posterior parameter for the Normal-Gamma, assuming that the observed values came from a Normal model from which the prior distribuition for the mean and the precision have joint distribuition Normal-Gamma
}
