% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_normal.R
\name{update_NG_alt}
\alias{update_NG_alt}
\title{update_NG_alt}
\usage{
update_NG_alt(conj_prior, ft, Qt, y, parms = list())
}
\arguments{
\item{conj_prior}{list: A vector containing the parameters of the Normal-Gamma (mu0,c0,alpha,beta). Not used in the alternative method.}

\item{ft}{vector: A vector representing the means from the normal distribution.}

\item{Qt}{matrix: A matrix representing the covariance matrix of the normal distribution.}

\item{y}{vector: A vector containing the observations.}

\item{parms}{list: A list of extra known parameters of the distribuition. Not used in this kernel.}
}
\value{
The parameters of the posterior distribution.
}
\description{
Calculate posterior parameter for the Normal-Gamma, assuming that the observed values came from a Normal model from which the prior distribuition for the mean and the precision have joint distribuition Normal-Gamma
}
