% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_normal.R
\name{update_Normal_dummy}
\alias{update_Normal_dummy}
\title{update_Normal_dummy}
\usage{
update_Normal_dummy(conj_prior, ft, Qt, y, parms)
}
\arguments{
\item{conj_prior}{list: A vector containing the concentration parameters of the Normal.}

\item{ft}{vector: A vector representing the means from the normal distribution. Not used in the default method.}

\item{Qt}{matrix: A matrix representing the covariance matrix of the normal distribution. Not used in the default method.}

\item{y}{vector: A vector containing the observations.}

\item{parms}{list: A list of extra known parameters of the distribuition. For this kernel, parms should containg the covariance matrix parameter (Sigma) for the observational Normal model.}
}
\value{
The parameters of the posterior distribution.
}
\description{
Calculate posterior parameter for the Normal, assuming that the observed values came from a Normal model from which the covariance is known and the prior distribuition for the mean vector have Normal distribuition
}
