% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{analytic_filter}
\alias{analytic_filter}
\title{normal_gamma_fit}
\usage{
analytic_filter(
  outcome,
  m0 = 0,
  C0 = 1,
  FF,
  G,
  D,
  W,
  offset,
  family,
  parms = list()
)
}
\arguments{
\item{outcome}{Matrix: The observed data. It's dimension shoulb be T x m, where T is the length of the time series and m is the number of outcomes at each time.}

\item{m0}{Vector: The prior mean for the latent vector.}

\item{C0}{Matrix: The prior covariance matrix for the latent vector.}

\item{FF}{Array: A 3D-array containing the regression matrix for each time. It's dimension should be n x m x T, where n is the number of latent variables, m is the number of outcomes in the model and T is the time series length.}

\item{G}{Matrix: The state evolution matrix.}

\item{D}{Array: A 3D-array containing the discount factor matrix for each time. It's dimension should be n x n x T, where n is the number of latent variables and T is the time series length.}

\item{W}{Array: A 3D-array containing the covariance matrix of the noise for each time. It's dimension should be the same as D.}

\item{offset}{Matrix: The offset of the model. It's dimension should be the same as outcome.}

\item{family}{function: the method for filtering.}

\item{parms}{list: a list contaning extra arguments. In this model, extra parameters are not used.}
}
\value{
A list containing the following values:
\itemize{
   \item mt Matrix: The filtered mean of the latent variables for each time. Dimensions are n x T.
   \item Ct Array: A 3D-array containing the filtered covariance matrix of the latent variable for each time. Dimensions are n x n x T.
   \item ft Matrix: The one-step-ahead linear predictor for each time. Dimensions are m x T.
   \item qt Array: A 3D-array containing the one-step-ahead covariance matrix for the linear predictor for each time. Dimensions are m x T.
   \item FF Array: The same as the argument (same values).
   \item G Matrix: The same as the argument (same values).
   \item D Array: The same as the argument (same values).
   \item W Array: The same as the argument (same values).
   \item offset Vector: The same as the argument (same values).
   \item outcome Matrix: The same as the argument outcome (same values).
   \item parms: The same as the argument.
}
}
\description{
Fit the  normal model given the observed value and the model parameters.
}
