% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{analytic_filter}
\alias{analytic_filter}
\title{analytic_filter}
\usage{
analytic_filter(
  outcomes,
  m0 = 0,
  C0 = 1,
  FF,
  G,
  D,
  W,
  p_monit = NA,
  c_monit = 1
)
}
\arguments{
\item{outcomes}{List: The observed data. It should contain objects of the class dlm_distr.}

\item{m0}{Vector: The prior mean for the latent vector.}

\item{C0}{Matrix: The prior covariance matrix for the latent vector.}

\item{FF}{Array: A 3D-array containing the regression matrix for each time. It's dimension should be n x m x T, where n is the number of latent variables, m is the number of outcomes in the model and T is the time series length.}

\item{G}{Array: A 3D-array containing the state evolution matrix at each time.}

\item{D}{Array: A 3D-array containing the discount factor matrix for each time. It's dimension should be n x n x T, where n is the number of latent variables and T is the time series length.}

\item{W}{Array: A 3D-array containing the covariance matrix of the noise for each time. It's dimension should be the same as D.}

\item{p_monit}{numeric (optional): The prior probability of changes in the latent space variables that are not part of it's dinamic.}

\item{c_monit}{numeric (optional, if p_monit is not specified): The relative cost of false alarm in the monitoring compared to the cost of not detecting anormalities.}
}
\value{
A list containing the following values:
\itemize{
   \item mt Matrix: The filtered mean of the latent variables for each time. Dimensions are n x T.
   \item Ct Array: A 3D-array containing the filtered covariance matrix of the latent variable for each time. Dimensions are n x n x T.
   \item at Matrix: The one-step-ahead mean of the latent variables at each time. Dimensions are n x T.
   \item Rt Array: A 3D-array containing the one-step-ahead covariance matrix for latent variables at each time. Dimensions are n x n x T.
   \item ft Matrix: The one-step-ahead linear predictor for each time. Dimensions are m x T.
   \item Qt Array: A 3D-array containing the one-step-ahead covariance matrix for the linear predictor for each time. Dimensions are m x m x T.
   \item FF Array: The same as the argument (same values).
   \item G Matrix: The same as the argument (same values).
   \item D Array: The same as the argument (same values) when there is no monitoring. When monitoring for anormalities, the value in times where anormalities were detected is increased.
   \item W Array: The same as the argument (same values).
   \item outcome List: The same as the argument outcome (same values).
   \item var_names Vector: The names of the linear predictors.
}
}
\description{
Fit the model given the observed value and the model parameters.
}
