% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_normal.R
\name{convert_dummy_post}
\alias{convert_dummy_post}
\title{convert_dummy_post}
\usage{
convert_dummy_post(conj_prior, parms)
}
\arguments{
\item{conj_prior}{list: A vector containing the parameters of the Normal}

\item{parms}{list: A list of extra known parameters of the distribution. Not used in this function.}
}
\value{
The parameters of the Normal distribution of the linear predictor.
}
\description{
Calculate the parameters of the Normal that best approximates the given Normal distribution.
The approximation is the best in the sense that it minimizes the KL divergence from the Normal to the Normal
This funciton is a dummy used for compatibility between kernels.
}
\seealso{
Other {auxiliary functions for a Normal outcome}: 
\code{\link{convert_NG_Normal}()},
\code{\link{convert_dummy_prior}()},
\code{\link{convert_multi_NG_Normal}()},
\code{\link{multi_normal_gamma_pred}()},
\code{\link{normal_pred}()},
\code{\link{update_NG_alt}()},
\code{\link{update_Normal_dummy}()},
\code{\link{update_multi_NG_correl}()}
}
\concept{{auxiliary functions for a Normal outcome}}
\keyword{internal}
