% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{dlm_sampling}
\alias{dlm_sampling}
\title{dlm_sampling}
\usage{
dlm_sampling(model, sample_size)
}
\arguments{
\item{model}{fitted_dlm: A fitted model from which to sample.}

\item{sample_size}{integer: The number of samples to draw.}
}
\value{
A list containing the following values:
\itemize{
   \item mt Array: An array containing the samples of latent states. Dimensions are n x T x sample_size, where n is the number of latent variable in the model and T is the number of observed values.
   \item ft Array: An array containing the samples of linear predictors. Dimensions are m x T x sample_size, where m is the number of linear predictors in the model and T is the number of observed values.
   \item param Array: An array containing the samples of the parameters of the observational model. Dimensions are k x T x sample_size, where k is the number of parameters in the observational model and T is the number of observed values.
}
}
\description{
dlm_sampling
}
\examples{

T <- 200
w <- (200 / 50) * 2 * pi
S <- exp(2 * (sin(w * 1:T / T))) * exp(-20 * (1:T / T) * ((1:T / T) - 1)) / 20
mu <- 20 * 1:T / T
outcome <- rnorm(T, mu, sqrt(1 / S))

level <- polynomial_block(
  order = 2,
  values = c(1, 0),
  D = 1 / 1,
  by_time = FALSE
)
variance1 <- polynomial_block(
  order = 3,
  values = c(0, 1),
  D = 1 / 1,
  C0 = diag(c(1, 0.1, 0.01)),
  by_time = FALSE
)
variance2 <- harmonic_block(
  period = 50,
  values = c(0, 1),
  D = 1 / 1,
  by_time = FALSE
)

fitted_data <- fit_model(level, variance1, variance2,
  outcome = outcome,
  family = "normal_gamma"
)

show_fit(fitted_data, smooth = TRUE)$plot

sample <- dlm_sampling(fitted_data, 2000)
}
