% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{fitted.values.fitted_dlm}
\alias{fitted.values.fitted_dlm}
\title{fitted.values.fitted_dlm}
\usage{
\method{fitted}{values.fitted_dlm}(fitted_dlm, smooth = TRUE, t_offset = 0, pred_cred = 0.95)
}
\description{
fitted.values method for class fitted_dlm
}
