% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma_kernel.R
\name{gamma_pred}
\alias{gamma_pred}
\title{gamma_pred}
\usage{
gamma_pred(conj_param, outcome = NULL, parms = list(), pred_cred = 0.95)
}
\arguments{
\item{conj_param}{List or data.frame: The parameters of the conjugated distribuitions of the linear predictor.}

\item{outcome}{Vector or matrix (optional): The observed values at the current time. Not used in this function.}

\item{parms}{List: A list of extra parameters for the model. For this function, it must contain the shape parameter phi of the observational model.}

\item{pred_cred}{Numeric: the desired credibility for the credibility interval.}
}
\value{
A list containing the following values:
\itemize{
   \item pred vector/matrix: the mean of the predictive distribuition of a next observation. Same type and shape as the parameter in model.
   \item var.pred vector/matrix: the variance of the predictive distribuition of a next observation. Same type and shape as the parameter in model.
   \item icl.pred vector/matrix: the percentile of 100*((1-pred_cred)/2)% of the predictive distribuition of a next observation. Same type and shape as the parameter in model.
   \item icu.pred vector/matrix: the percentile of 100*(1-(1-pred_cred)/2)% of the predictive distribuition of a next observation. Same type and shape as the parameter in model.
}
}
\description{
Calculate the values for the predictive distribuition given the values of the parameter of the conjugated distribuition of the linear predictor.
The data is assumed to have Gamma distribuition with known shape parameter phi and it's mean having distribuition Inverse-Gamma with shape parameter a e rate parameter b.
In this scenario, the marginal distribuition of the data is Beta prime with parameters phi, alpha, beta / phi.
}
\examples{

conj_param <- data.frame(
  "alpha" = c(1:3),
  "beta" = c(3:1)
)

gamma_pred(conj_param, parms = list("phi" = 1))
}
