% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma_kernel.R
\name{gamma_pred}
\alias{gamma_pred}
\title{gamma_pred}
\usage{
gamma_pred(model, pred_cred = 0.95)
}
\arguments{
\item{model}{List: contaning the parameters of the conjugated distribuition for the linear predictor (may be based on the prior, filtered or smoothed distribuition).}

\item{pred_cred}{Numeric: the desired credibility for the credibility interval.}
}
\value{
A list containing the following values:
\itemize{
   \item pred vector/matrix: the mean of the predictive distribuition of a next observation. Same type and shape as the parameter in model.
   \item var.pred vector/matrix: the variance of the predictive distribuition of a next observation. Same type and shape as the parameter in model.
   \item icl.pred vector/matrix: the percentile of 100*((1-pred_cred)/2)% of the predictive distribuition of a next observation. Same type and shape as the parameter in model.
   \item icu.pred vector/matrix: the percentile of 100*(1-(1-pred_cred)/2)% of the predictive distribuition of a next observation. Same type and shape as the parameter in model.
   \item log.like vector/matrix: the log likelyhood of the observed data based on the predictive distribuition.
}
}
\description{
Calculate the values for the predictive distribuition given the values of the parameter of the conjugated distribuition of the linear predictor.
It's worth noting that the credibility interval is calculated as +-2 standard deviations.
}
\examples{
# A fitted model shoulb be used as argument, but you can also pass only the parameter themselves.

model <- list(
  "outcome" = 2,
  "tau0" = 3,
  "tau1" = 4,
  "parms" = list("alpha" = 1)
)

gamma_pred(model)
}
