% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multnom_kernel.R
\name{multnom_pred}
\alias{multnom_pred}
\title{multnom_pred}
\usage{
multnom_pred(conj_param, outcome, parms = list(), pred_cred = 0.95)
}
\arguments{
\item{pred_cred}{Numeric: the desired credibility for the credibility interval}

\item{model}{List: contaning the parameters of the conjugated distribuition for the linear predictor (may be based on the prior, filtered or smoothed distribuition).}
}
\value{
A list containing the following values:
\itemize{
   \item pred vector: the mean of the predictive distribuition of a next observation. Same type and shape as the parameter in model.
   \item var.pred vector: the variance of the predictive distribuition of a next observation. Same type and shape as the parameter in model.
   \item icl.pred vector: the percentile of 100*((1-pred_cred)/2)% of the predictive distribuition of a next observation. Same type and shape as the parameter in model.
   \item icu.pred vector: the percentile of 100*(1-(1-pred_cred)/2)% of the predictive distribuition of a next observation. Same type and shape as the parameter in model.
   \item log.like vector/matrix: the log likelyhood of the observed data based on the predictive distribuition.
}
}
\description{
This function is a temporary placeholder.
Calculate the values for the predictive distribuition given the values of the parameter of the conjugated distribuition of the linear predictor.
PLACE PREDICTIVE DISTRIBUITION.
}
\examples{
# MUST BE MADE.
# A fitted model shoulb be used as argument, but you can also pass only the parameter themselves.

model <- list(
  "alpha" = c(10, 2, 6),
  "outcome" = c(3, 2, 5)
)

multnom_pred(model)
}
