% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_kernel.R
\name{normal_fit}
\alias{normal_fit}
\title{normal_fit}
\usage{
normal_fit(
  outcome,
  m0 = 0,
  C0 = 1,
  FF,
  G,
  D,
  W,
  offset = outcome * 0,
  pred_cred = 0.95,
  parms = list(),
  kernel_filter = normal_gamma_filter
)
}
\arguments{
\item{outcome}{Matrix: The observed data. It's dimension shoulb be T x m, where T is the length of the time series and m is the number of outcomes at each time.}

\item{m0}{Vector: The prior mean for the latent vector.}

\item{C0}{Matrix: The prior covariance matrix for the latent vector.}

\item{FF}{Array: A 3D-array containing the regression matrix for each time. It's dimension should be n x m x T, where n is the number of latent variables, m is the number of outcomes in the model and T is the time series length.}

\item{G}{Matrix: The state evolution matrix.}

\item{D}{Array: A 3D-array containing the discount factor matrix for each time. It's dimension should be n x n x T, where n is the number of latent variables and T is the time series length.}

\item{W}{Array: A 3D-array containing the covariance matrix of the noise for each time. It's dimension should be the same as D.}

\item{offset}{Matrix: The offset of the model. It's dimension should be the same as outcome.}

\item{pred_cred}{Numeric: the desired credibility for the credibility interval.}

\item{parms}{list: a list contaning extra arguments. In this model, sigma2 (covariance matrix for the noise associated withy) should be passed.}

\item{kernel_filter}{function: the method for filtering.}
}
\value{
A list containing the following values:
\itemize{
   \item mt Matrix: The filtered mean of the latent variables for each time. Dimensions are n x T.
   \item Ct Array: A 3D-array containing the filtered covariance matrix of the latent variable for each time. Dimensions are n x n x T.
   \item ft Matrix: The one-step-ahead linear predictor for each time. Dimensions are m x T.
   \item qt Array: A 3D-array containing the one-step-ahead covariance matrix for the linear predictor for each time. Dimensions are m x T.
   \item FF Array: The same as the argument (same values).
   \item G Matrix: The same as the argument (same values).
   \item D Array: The same as the argument (same values).
   \item W Array: The same as the argument (same values).
   \item pred Matrix: The one-step-ahead predictions for each time. Dimensions are m x T.
   \item var.pred Matrix: The variance for the one-step-ahead predictions for each time. Dimensions are m x T. Note that, in the multivariate Poisson case, the series are supossed independent, so, in particular, they are uncorrelated.
   \item icl.pred Matrix: The lower credibility interval for the prediction at each time. Dimensions are m x T.
   \item icu.pred Matrix: The upper credibility interval for the prediction at each time. Dimensions are m x T.
   \item pred_cred Numeric: Deprecated
   \item offset Vector: The same as the argument (same values).
   \item log.vero Matrix: log likelyhood for the filtered parameters based on the one-step ahead prediction.
   \item outcome Matrix: The same as the argument outcome (same values).
   \item parms: The same as the argument.
}
}
\description{
Fit the  normal model given the observed value and the model parameters.
}
\examples{
# Not ideal way: should use fit_model function.
T <- 200
mean <- 0
var <- 1
outcome <- matrix(rnorm(T) * sqrt(var) + mean)
m0 <- c(0, 0)
C0 <- diag(c(1, 1))
G <- matrix(c(1, 0, 1, 1), 2, 2)
FF <- array(matrix(c(1, 0), 2, 1), c(2, 1, T))
D <- array(diag(c(0, 0)), c(2, 2, T)) + 1
W <- array(diag(c(0, 0)), c(2, 2, T))

fitted_data <- normal_fit(outcome = outcome, m0 = m0, C0 = C0, FF = FF, G = G, D = D, W = W, pred_cred = 0.95, parms = list("sigma2" = 1))

plot(outcome)
lines(fitted_data$pred[1, ])
}
