% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helper.R
\name{show_fit}
\alias{show_fit}
\title{show_fit}
\usage{
show_fit(
  model,
  IC_prob = 0.95,
  smooth = TRUE,
  dinamic = TRUE,
  t_offset = 0,
  labels = NULL
)
}
\arguments{
\item{labels}{}
}
\description{
show_fit
}
