% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helper.R
\name{show_fit}
\alias{show_fit}
\title{show_fit}
\usage{
show_fit(
  model,
  pred_cred = 0.95,
  smooth = TRUE,
  dynamic_plot = TRUE,
  t_offset = 0,
  labels = NULL
)
}
\arguments{
\item{model}{<undefined class> or list: A fitted GDLM model.}

\item{pred_cred}{Numeric: The credibility value for the credibility interval.}

\item{smooth}{Bool: A flag indicating if the smoothed should be used. If false, the filtered distribuition will be used.}

\item{t_offset}{Integer: A integer with the amount of steps ahead should be used for prediciton. Only used if smooth is false.}

\item{labels}{Character: A vector containg the names for each time series.}

\item{dynamic}{Bool: A flag indicating if the created plot should be dynamic.}
}
\value{
A list containg:
\itemize{
   \item data tibble object: A data frame containing the observations, predictions and credibility intervals at each time.
   \item plot ggplot or plotly object: A plot showing the predictive mean and credibility interval with the observed data.
}
}
\description{
Calculate the preditive mean and some quantiles for the observed data and show a plot.
}
\examples{
T <- 200
w <- (200 / 40) * 2 * pi
outcome <- rpois(T, 20 * (sin(w * 1:T / T) + 2))

level <- polynomial_block(order = 1, values = 1, D = 1 / 0.95)
season <- harmonic_block(period = 40, values = 1, D = 1 / 0.98)

fitted_data <- fit_model(level, season, outcome = outcome, family = "Poisson")
show_fit(fitted_data, smooth = TRUE)$plot
}
