% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_dirichlet.R
\name{update_Dirichlet}
\alias{update_Dirichlet}
\title{update_Dirichlet}
\usage{
update_Dirichlet(conj_prior, ft, Qt, y, parms)
}
\arguments{
\item{conj_prior}{list: A vector containing the parameters of the Inverse-Gamma (alpha,beta). Not used in the alternative method.}

\item{ft}{vector: A vector representing the means from the prior distribution.}

\item{Qt}{matrix: A matrix representing the covariance matrix of the prior distribution.}

\item{y}{vector: A vector containing the observations.}

\item{parms}{list: A list of extra known parameters of the distribuition. For this kernel, parms should containg the shape parameter (phi) for the observational gamma model.}
}
\value{
The parameters of the posterior distribution.
}
\description{
Dirichlet
}
