% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_kernel.R
\name{update_Normal_dummy}
\alias{update_Normal_dummy}
\title{update_Normal_dummy}
\usage{
update_Normal_dummy(conj_prior, y, parms)
}
\arguments{
\item{conj_prior}{list: A vector containing the concentration parameters of the Normal.}

\item{y}{vector: A vector containing the observations.}

\item{parms}{list: A list of extra known parameters of the distribuition. For this kernel, parms should containg the covariance matrix parameter (Sigma) for the observational Normal model.}
}
\description{
Calculate posterior parameter for the Normal, assuming that the observed values came from a Normal model from which the covariance is known and the prior distribuition for the mean vector have Normal distribuition
}
